query_node: fix implicit AND handling
We can only add an implicit AND after we've successfully parsed the next
token and before we're about to add it to the suffix expression list.
Otherwise it's possible that we already added an implicit AND but the
parsing of the next token fails (e.g. due to an invalid format), then we
have an AND which applies to a wrong or absent operand.