filter: fix endless loop in recursive macro definitions
Up until now it was possible for a macro to call itself, directly and
indirectly, which effectively resulted in an endless loop while
expanding the macro.

To fix that we now use a stack to detect nested macros and ignore them.

On top of that we now also parse macros with the existing query parser.
This makes parsing much more robust and fixes some issues the macro
expansion had with escape sequences etc.