add config for showing dialog on deletion
Adds a config option as a sibling of the existing option that chooses
whether or not to show a dialog when files fail to open. This toggles
showing a dialog when deleting more than 20 items (for now), and also
disables the informational dialog that tells the user that the database
view will not update and only the thumbnails will. This information has
been copied into the help section.